requirement.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.7
django-cors-headers==3.3.0
django-phonenumber-field==4.0.0
django-rest-knox==4.1.0
django-sendsms==0.4
django-twilio==0.13.0.post1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
phonenumbers==8.12.5
Pillow==7.1.2
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
twilio==6.41.0
urllib3==1.25.9
